아마 encoding 때문에 json parsing이 안되는거 같다..

+++ air_condition/src/http.h
@@ -4,8 +4,8 @@
 #include "ArduinoJson.h"
 
 struct FineParticle {
-  float pm25;
-  float pm10;
+  double pm25;
+  double pm10;
   String error;
 };
